Revert "Integrate write endpoints in UDS activity"
This reverts commit 26e3d3d158847825d9a990cb8f587d64bf16bd6c.
