+++ src/android_app/CarsData/app/src/main/java/com/poc/p_couds/HandleNetworkReq.kt
@@ -1,6 +1,5 @@
 package com.poc.p_couds
 
-import com.poc.p_couds.models.IApiService
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
