--- src/android_app/CarsData/app/src/main/java/com/poc/p_couds/pojo/WriteBatteryChargeStateDataClass.kt
@@ -1,5 +1,0 @@
-package com.example.jetpackgettingstarted.pojos
-
-data class WriteBatteryChargeStateDataClass(
-    val battery_state_of_charge: Any
-)
