--- src/android_app/CarsData/app/src/main/java/com/poc/p_couds/pojo/WriteBatteryResponseDataClass.kt
@@ -1,5 +1,0 @@
-package com.poc.p_couds.pojo
-
-data class WriteBatteryResponseDataClass(
-    val response: Any
-)
