--- src/android_app/CarsData/app/src/main/java/com/poc/p_couds/pojo/WriteDoorsPassengerLockDataClass.kt
@@ -1,5 +1,0 @@
-package com.poc.p_couds.pojo
-
-data class WriteDoorsPassengerLockDataClass(
-    val passenger_lock: Any
-)
