--- src/android_app/CarsData/app/src/main/java/com/poc/p_couds/pojo/WriteEngineIntakeAirTempDataClass.kt
@@ -1,5 +1,0 @@
-package com.poc.p_couds.pojo
-
-data class WriteEngineIntakeAirTempDataClass(
-    val intake_air_temperature: Any
-)
