--- src/android_app/CarsData/app/src/main/java/com/poc/p_couds/pojo/WriteHvacAmbientAirTempDataClass.kt
@@ -1,5 +1,0 @@
-package com.poc.p_couds.pojo
-
-data class WriteHvacAmbientAirTempDataClass(
-    val ambient_air_temperature: Any
-)
