--- src/android_app/CarsData/app/src/main/java/com/poc/p_couds/pojo/WriteHvacCabinTempDataClass.kt
@@ -1,5 +1,0 @@
-package com.poc.p_couds.pojo
-
-data class WriteHvacCabinTempDataClass(
-    val cabin_temperature: Any
-)
