--- src/android_app/CarsData/app/src/main/java/com/poc/p_couds/pojo/WriteHvacCabinTempDriverSetDataClass.kt
@@ -1,5 +1,0 @@
-package com.poc.p_couds.pojo
-
-data class WriteHvacCabinTempDriverSetDataClass(
-    val cabin_temperature_driver_set: Any
-)
