--- src/android_app/CarsData/app/src/main/java/com/poc/p_couds/pojo/WriteHvacFanSpeedDataClass.kt
@@ -1,5 +1,0 @@
-package com.poc.p_couds.pojo
-
-data class WriteHvacFanSpeedDataClass(
-    val fan_speed: Any
-)
