--- src/android_app/CarsData/app/src/main/java/com/poc/p_couds/pojo/WriteHvacModesDataClass.kt
@@ -1,5 +1,0 @@
-package com.poc.p_couds.pojo
-
-data class WriteHvacModesDataClass(
-    val hvac_modes: Any
-)
